+++ library-src/Sound/Punctual/Parser.hs
@@ -17,6 +17,8 @@ import Control.Monad.Except
 import Data.Maybe
 import Control.Applicative
 import Data.Time
+import Data.Bifunctor
+import Data.Char (isSpace)
 
 import Sound.Punctual.Extent
 import Sound.Punctual.Graph
@@ -33,7 +35,7 @@ parse eTime x = do
   let (x',pragmas) = extractPragmas x
   if (elem "glsl" pragmas) then do
     return $ (emptyProgram eTime) { directGLSL = Just x' }
-  else parseProgram eTime $ T.unpack x'
+  else first errorMessage $ parseProgram eTime $ T.unpack x'
 
 extractPragmas :: Text -> (Text,[Text])
 extractPragmas t = (newText,pragmas)
@@ -44,19 +46,25 @@ extractPragmas t = (newText,pragmas)
     newText = T.unlines $ fmap fst xs
     pragmas = concat $ fmap snd xs
 
-reformatProgramAsList :: String -> String
-reformatProgramAsList x =
-  let x' = intercalate "," $ fmap (++ " _0") $ splitOn ";" x
-  in "[" ++ x' ++ "\n]"
 
-parseProgram :: UTCTime -> String -> Either String Program
+errorMessage :: (Span,Text) -> String
+errorMessage (s,m) = show s ++ " " ++ T.unpack m
+
+parseProgram :: UTCTime -> String -> Either (Span,Text) Program
 parseProgram eTime x = do
-  (p,st) <- parseHaskellish (program eTime) emptyParserState $ reformatProgramAsList x
+  (p,st) <- parseWithModeAndRun haskellSrcExtsParseMode (program eTime) emptyParserState $ reformatProgramAsList $ removeComments x
   return $ p {
     textureSet = textureRefs st,
     programNeedsAudioInputAnalysis = audioInputAnalysis st,
     programNeedsAudioOutputAnalysis = audioOutputAnalysis st
   }
+
+reformatProgramAsList :: String -> String
+reformatProgramAsList x =
+  let x' = splitOn ";" x
+      x'' = Data.List.filter (\y -> length (dropWhile isSpace y) > 0) x'
+      x''' = intercalate "," x''
+  in "[" ++ x''' ++ "\n]"
 
 
 type Identifier = String
@@ -85,13 +93,6 @@ emptyParserState = ParserState {
   }
 
 type H = Haskellish ParserState
-
-parseHaskellish :: H a -> ParserState -> String -> Either String (a,ParserState)
-parseHaskellish p st x = (parseResultToEither $ parseWithMode haskellSrcExtsParseMode x) >>= runHaskellish p st
-
-parseResultToEither :: ParseResult a -> Either String a
-parseResultToEither (ParseOk x) = Right x
-parseResultToEither (ParseFailed _ s) = Left s
 
 haskellSrcExtsParseMode :: ParseMode
 haskellSrcExtsParseMode = defaultParseMode {
@@ -139,20 +140,11 @@ haskellSrcExtsParseMode = defaultParseMode {
     }
 
 
-_0Arg :: H a -> H a
-_0Arg p = p <|> fmap fst (functionApplication p $ reserved "_0")
-
 program :: UTCTime -> H Program
 program eTime = do
-  xs <- concat <$> list actionOr_0
+  xs <- list action
   let xs' = Data.List.filter (not . Data.List.null . Sound.Punctual.Action.outputs) xs
   return $ (emptyProgram eTime) { actions = IntMap.fromList $ zip [0..] xs' }
-
-actionOr_0 :: H [Action]
-actionOr_0 = _0Arg $ asum [
-  reserved "_0" >> return [],
-  fmap pure action
-  ]
 
 action :: H Action
 action = asum [
@@ -188,7 +180,7 @@ number = asum [
   ]
 
 duration :: H Duration
-duration = _0Arg $ asum [
+duration = asum [
   Seconds <$> rationalOrInteger,
   reverseApplication number (reserved "s" >> return Seconds),
   reverseApplication number (reserved "ms" >> return (\x -> Seconds $ x/1000.0)),
@@ -196,13 +188,13 @@ duration = _0Arg $ asum [
   ]
 
 defTime :: H DefTime
-defTime = _0Arg $ asum [
+defTime = asum [
   (\(x,y) -> Quant x y) <$> Language.Haskellish.tuple number duration,
   After <$> duration
   ]
 
 outputs :: H [Output]
-outputs = _0Arg $ asum [
+outputs = asum [
   concat <$> list Sound.Punctual.Parser.outputs,
   ((:[]) . Panned . realToFrac) <$> rationalOrInteger,
   reserved "audio" >> return [Splay],
@@ -236,7 +228,7 @@ debugExp h = do
   throwError $ T.pack $ show e
 
 graph :: H Graph
-graph = _0Arg $ asum [
+graph = asum [
   identifiedGraph,
   definitions1H,
   reverseApplication graph (reserved "m" >> return MidiCps),
@@ -281,7 +273,7 @@ ifThenElseParser = do
   return $ IfThenElse a b c
 
 graph2 :: H (Graph -> Graph)
-graph2 = _0Arg $ asum [
+graph2 = asum [
   reserved "bipolar" >> return Bipolar,
   reserved "unipolar" >> return Unipolar,
   reserved "sin" >> return Sin,
@@ -412,8 +404,30 @@ textureRef = (do
   return t
   ) <?> "expected texture URL"
 
+
 multiSeries :: H Graph
-multiSeries = (reserved "..." >> return f) <*> i <*> i
+multiSeries = multiSeries0 <|> multiSeries1 <|> multiSeriesDeprecated
+
+multiSeries0 :: H Graph
+multiSeries0 = do
+  let i = fromIntegral <$> integer
+  (a,b) <- Language.Haskellish.enumFromTo i i
+  return $ Multi $ fmap Constant $ Data.List.take 64 [a .. b]
+
+multiSeries1 :: H Graph
+multiSeries1 = do
+  (a,b,c) <- Language.Haskellish.enumFromThenTo double double double
+  return $ Multi $ fmap Constant $ Data.List.take 64 $ multiSeries1' a b c
+
+multiSeries1' :: Double -> Double -> Double -> [Double]
+multiSeries1' a b c
+  | (b > a) && (a > c) = []
+  | (b < a) && (a < c) = []
+  | (b == a) = []
+  | otherwise = a : multiSeries1' b (b+b-a) c
+
+multiSeriesDeprecated :: H Graph
+multiSeriesDeprecated = (reserved "..." >> return f) <*> i <*> i
   where
-    f x y = Multi $ fmap Constant [x .. y]
+    f x y = Multi $ fmap Constant $ Data.List.take 64 [x .. y]
     i = fromIntegral <$> integer