+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 - significant refactor of fragment shader generation
 - many changes to multi-channel expansion behaviours
 
+0.3.4:
+- support for multichannel audio
+
+0.3.3:
+
+- support for Haskell-style multi-channel ranges of integer constants [1..6] and float/double constants [0.1,0.2 .. 0.9]
+
 0.3.2.1:
+
 - bug fix relating to line, iline
 
 0.3.2:
@@ -14,7 +22,6 @@
 0.3.1:
 
 - new: added 'delay' for delay lines (audio only)
-
 
 0.3: